+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Look up WC kubeconfig based on Cluster name instead of OpenStackCluster name.
+- Update changed DNS record values.
+- Use name from Cluster instead of OpenStackCluster.
+- Skip WC ingress IP lookup during cluster deletion.
+- Cache WC k8s client in scope.
+- Create WC k8s client when needed rather than on every reconciliation loop.
 
 ## [0.1.0] - 2021-12-15